+++ src/mantle/events/class-dispatcher.php
@@ -14,12 +14,12 @@ use Mantle\Contracts\Container;
 use Mantle\Contracts\Events\Dispatcher as Dispatcher_Contract;
 use Mantle\Support\Arr;
 use Mantle\Support\Str;
+use RuntimeException;
 
 /**
  * Event Dispatcher
  *
  * @todo Add queued event listeners.
- * @todo Add wildcard listeners.
  */
 class Dispatcher implements Dispatcher_Contract {
 	use WordPress_Action;
@@ -37,11 +37,18 @@ class Dispatcher implements Dispatcher_Contract {
 	protected Container $container;
 
 	/**
-	 * The queue resolver instance.
+	 * Wildcard listeners.
 	 *
-	 * @var callable
+	 * @var array<string, array<string|callable>>
 	 */
-	protected $queue_resolver;
+	protected array $wildcard_listeners = [];
+
+	/**
+	 * Wildcard listener lookup cache.
+	 *
+	 * @var array<string, array<string|callable>>
+	 */
+	protected array $wildcard_cache = [];
 
 	/**
 	 * Create a new event dispatcher instance.
@@ -58,17 +65,23 @@ class Dispatcher implements Dispatcher_Contract {
 	 * @todo Add wildcard listeners.
 	 *
 	 * @param string|string[] $events Event(s) to listen to.
-	 * @param mixed        $listener Listener to register.
+	 * @param string|callable        $listener Listener to register.
 	 * @param int          $priority Event priority.
 	 * @param  \Closure|string $listener Listener callback.
 	 */
-	public function listen( $events, $listener, int $priority = 10 ): void {
+	public function listen( string|array $events, string|callable $listener, int $priority = 10 ): void {
 		foreach ( (array) $events as $event ) {
+			if ( str_contains( $event, '*' ) ) {
+				$this->setup_wildcard_listener( $event, $listener );
+
+				continue;
+			}
+
 			add_action(
 				$event,
 				$this->make_listener( $listener ),
 				$priority,
-				PHP_INT_MAX,
+				999,
 			);
 		}
 	}
@@ -78,8 +91,17 @@ class Dispatcher implements Dispatcher_Contract {
 	 *
 	 * @param  string $event_name Event name.
 	 */
-	public function has_listeners( $event_name ): bool {
-		return has_filter( $event_name );
+	public function has_listeners( string $event_name ): bool {
+		return has_filter( $event_name ) || $this->has_wildcard_listeners( $event_name );
+	}
+
+	/**
+	 * Determine if a given event has wildcard listeners.
+	 *
+	 * @param  string $event_name Event name.
+	 */
+	public function has_wildcard_listeners( string $event_name ): bool {
+		return ! empty( $this->get_wildcard_listeners( $event_name ) );
 	}
 
 	/**
@@ -87,7 +109,7 @@ class Dispatcher implements Dispatcher_Contract {
 	 *
 	 * @param  object|string $subscriber
 	 */
-	public function subscribe( $subscriber ): void {
+	public function subscribe( object|string $subscriber ): void {
 		$subscriber = $this->resolve_subscriber( $subscriber );
 
 		$subscriber->subscribe( $this );
@@ -99,7 +121,7 @@ class Dispatcher implements Dispatcher_Contract {
 	 * @param  object|string $subscriber
 	 * @return mixed
 	 */
-	protected function resolve_subscriber( $subscriber ) {
+	protected function resolve_subscriber( object|string $subscriber ) {
 		if ( is_string( $subscriber ) ) {
 			return $this->container->make( $subscriber );
 		}
@@ -110,19 +132,28 @@ class Dispatcher implements Dispatcher_Contract {
 	/**
 	 * Fire an event and call the listeners.
 	 *
-	 * @todo Break out support for a filter.
+	 * @throws RuntimeException Thrown if the event is an object and payload is passed.
 	 *
 	 * @param  string|object $event Event name.
-	 * @param  mixed         $payload Event payload.
+	 * @param  mixed         ...$payload Event payload.
 	 */
-	public function dispatch( string|object $event, mixed $payload = [ null ] ): mixed {
-		[ $event, $payload ] = $this->parse_event_and_payload( $event, $payload );
-
-		if ( function_exists( 'apply_filters' ) ) {
-			return apply_filters( $event, ...$payload ); // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.DynamicHooknameFound
+	public function dispatch( string|object $event, mixed ...$payload ): mixed {
+		if ( is_object( $event ) && ! empty( $payload ) ) {
+			throw new RuntimeException( 'You cannot pass payload to an object event.' );
 		}
 
-		return null;
+		[ $event, $payload ] = $this->parse_event_and_payload( $event, $payload );
+
+		if ( ! function_exists( 'apply_filters' ) ) {
+			return null;
+		}
+
+		// Ensure there is a payload that is able to be passed to the filter.
+		if ( empty( $payload ) ) {
+			$payload[] = ''; // Mirror the default behavior of do_action.
+		}
+
+		return apply_filters( $event, ...$payload ); // phpcs:ignore WordPress.NamingConventions.PrefixAllGlobals.DynamicHooknameFound
 	}
 
 	/**
@@ -141,17 +172,27 @@ class Dispatcher implements Dispatcher_Contract {
 	}
 
 	/**
-	 * Get all of the listeners for a given event name.
+	 * Get all of the wildcard listeners for a given event name.
 	 *
 	 * @param  string $event_name
-	 * @return array<mixed>
+	 * @return array<string|callable>
 	 */
-	public function get_listeners( string $event_name ): array {
-		$listeners = $this->listeners[ $event_name ] ?? [];
+	public function get_wildcard_listeners( string $event_name ): array {
+		if ( isset( $this->wildcard_cache[ $event_name ] ) ) {
+			return $this->wildcard_cache[ $event_name ];
+		}
 
-		return class_exists( $event_name, false )
-			? $this->add_interface_listeners( $event_name, $listeners )
-			: $listeners;
+		$listeners = [];
+
+		foreach ( $this->wildcard_listeners as $pattern => $wildcard_listeners ) {
+			if ( Str::is( $pattern, $event_name ) ) {
+				$listeners = array_merge( $listeners, $wildcard_listeners );
+			}
+		}
+
+		$this->wildcard_cache[ $event_name ] = $listeners;
+
+		return $listeners;
 	}
 
 	/**
@@ -230,19 +271,96 @@ class Dispatcher implements Dispatcher_Contract {
 	/**
 	 * Remove a set of listeners from the dispatcher.
 	 *
-	 * @param string|object $event Event to remove.
-	 * @param callable|string $listener Listener to remove.
-	 * @param int $priority Priority of the listener.
+	 * @param string|object        $event Event to remove.
+	 * @param callable|string|null $listener Listener to remove.
+	 * @param int                  $priority Priority of the listener.
 	 */
-	public function forget( $event, $listener = null, int $priority = 10 ): void {
+	public function forget( string|object $event, string|callable|null $listener = null, int $priority = 10 ): void {
 		if ( is_object( $event ) ) {
 			$event = $event::class;
+		}
+
+		if ( str_contains( $event, '*' ) ) {
+			$this->forget_wildcard( $event, $listener );
+
+			return;
 		}
 
 		if ( null === $listener ) {
 			remove_all_filters( $event, $priority );
 		} else {
 			remove_filter( $event, $listener, $priority );
+		}
+	}
+
+	/**
+	 * Remove a wildcard listener from the dispatcher.
+	 *
+	 * @param string $event Event to remove.
+	 * @param string|callable|null $listener Listener to remove.
+	 */
+	public function forget_wildcard( string $event, string|callable|null $listener = null ): void {
+		if ( empty( $this->wildcard_listeners[ $event ] ) ) {
+			return;
+		}
+
+		$this->wildcard_cache = [];
+
+		if ( null === $listener ) {
+			unset( $this->wildcard_listeners[ $event ] );
+
+			return;
+		}
+
+		$this->wildcard_listeners[ $event ] = array_filter(
+			$this->wildcard_listeners[ $event ],
+			fn ( $value ) => $value !== $listener,
+		);
+
+		if ( empty( $this->wildcard_listeners[ $event ] ) ) {
+			unset( $this->wildcard_listeners[ $event ] );
+		}
+	}
+
+	/**
+	 * Setup a wildcard event listener.
+	 *
+	 * Registers a listener for the 'all' action which is fired for all hooks
+	 * which we can then use to find the appropriate listeners. Wildcard events
+	 * cannot have a priority.
+	 *
+	 * @param string   $event Event name to listen to with * wildcard.
+	 * @param callable $listener Listener to register.
+	 */
+	protected function setup_wildcard_listener( string $event, callable $listener ): void {
+		if ( function_exists( 'has_action' ) && ! has_action( 'all', [ $this, 'wildcard_listener_callback' ] ) ) {
+			add_action( 'all', [ $this, 'wildcard_listener_callback' ] );
+		}
+
+		$this->wildcard_cache = [];
+
+		$this->wildcard_listeners[ $event ][] = $listener;
+	}
+
+	/**
+	 * Callback for the wildcard listener.
+	 *
+	 * Hooked to the 'all' action to catch all events.
+	 *
+	 * @param string $hook Hook being fired.
+	 * @param mixed ...$args Arguments for the hook.
+	 */
+	public function wildcard_listener_callback( string $hook, mixed ...$args ): void {
+		foreach ( $this->wildcard_listeners as $pattern => $listeners ) {
+			if ( ! Str::is( $pattern, $hook ) ) {
+				continue;
+			}
+
+			foreach ( $listeners as $listener ) {
+				$callable = $this->create_action_callback( $listener );
+
+				$callable( $hook, ...$args );
+			}
 		}
 	}
 }