+++ CHANGELOG.md
@@ -21,6 +21,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added `Mantle\Support\Registration\Post_Type_Arguments` and
   `Mantle\Support\Registration\Taxonomy_Arguments` classes to allow for
   registering post types/taxonomies with custom arguments fluently.
+- Add wildcard event listener support to events.
+- Add support for more log drivers and channels.
+- Add `environment_mixed()` helper to return a `Mixed_Data` instance for an
+  environment variable.
 
 ### Changed